build: log when the task is locked
This improve the bitbake UI when there are tasks with
lockfiles waiting to run.

In such cases one or more tasks are waiting because
there is another one with the lock acquired.
The bitbake UI shows that all the tasks is running
which is not true and can confuse the user.

This patch will shows what are the task that have the lock
acquired.

Signed-off-by: Jose Quaresma <jose.quaresma@foundries.io>